Workaround for internal compiler error regarding some native function calls.
The function commented in this commit led errors like the following during compiling a Release-version of some UWP-app already:

> ICE: trying to add a local var with the same name, but different types. during [_RegisterClipboardFormat]

The CN1-fork simply commented those functions as well, so I followed that approach for now. The important thing is that I really only commented those functions which triggered the error until that error didn't occur anymore. So I did not simply commented all native functions by purpose, because currently it seems that some trigger that error, while some don't.

The problem is that currently another error is triggered during compilation, so I can't be sure if that is fixed somehow, the error of this commit happens again with the other functions not yet commented. Because of the second error, committing the current changes seems the best way to go to keep things independent.

This is related to https://github.com/ikvm-revived/ikvm/issues/1.