Workaround for code generating multiple `__value`-args.
The functions commented in this commit led to errors like the following during compiling a Release-version of some UWP-app:

> error CS0100: The parameter name '__value' is a duplicate"

The CN1-fork simply commented those functions as well, so I followed that approach for now. The important thing is that I really only commented those functions which triggered the error until that error didn't occur anymore. So I did not simply commented all native functions by purpose, because currently it seems that some trigger that error, while some don't. The difference most likely simply is the number of arguments, because on "__value" is not a problem, while two are.

This is related to https://github.com/ikvm-revived/ikvm/issues/2.