Made old usage of FileStream-CTOR available where possible.
Some FileStream-CTORs have not been ported from .NET Framework to .NET Standard or older versions of .NET Core, so using those has been removed. But things worked for .NET Framework and using that might be needed, so put the old code back in place and switched using "#if" where possible. That's not possible in Java-files, so kept them as they where for now.