Reimplementing java.net stack. Current implementation is based around the OpenJDK Windows implementation, which uses JNI to invoke Winsock. At some point somebody wrote a C# implementation of the Winsock API that calls into .NET. Which calls into Winsock. Insane. OpenJDK also has two divergent classes for all this: TwoStack and DualStack. Which are selected based on whether we're on pre-Vista or not.
Assuming .NET takes care of all of that, no reason for this. The .NET Socket API is reasonable similar to BSD. So instead, we're now using the Solaris implementation, and some C# that calls into .NET. There's not much too this so far. Just mapping some exception types and a small number of assumptions.

DatagramSocket is done. Needs good testing. PlainSocketImpl next.