Touch the reference export output to ensure mtimes are later. _CompileJavaResponseFile needs to happen after BeforeTargets, which is where the assemblyinfo.java files are generated.
For Nashorn, we need to merge in the generated class files (produced by nasgen). There are no source versions of these files. And their output is intermingled into the output of everything else. So, we'll just compile the class files, and not do a source compile.

Ideally this would be done differently. Nashorn should be split from IKVM.Java, and thus able to do a source compile using IKVM's versionof javac, and build nasgen, and execute nasgen. However, we can't merge the classes right now as the bootstrap class loader isn't coded to find anything other than IKVM.Java. Once we resolve that, we should be able to split nashorn (and langtools) out.