Remove WorkaroundAbstractMethodWidening Remove WorkaroundInterfaceFields Remove WorkaroundInterfacePrivateMethods Remove WorkaroundInterfaceStaticMethods
Each of these generate various compatibility code to make the IKVM IL compatible with pre-Roslyn C# compilers. We are a long way past that now. Enum types should be directly visible. Methods on fields should be accessible. Etc.

This changes the layout of produced .NET types, and is thus an API break.